fix(trusty-search): address 8 safety/correctness issues in prune command
Fixes all issues raised in the independent security review of PR #1786:

1. CRITICAL (config error propagation): malformed config.yaml now aborts
   --apply with an error instead of silently using empty protected_indexes.
   Dry-runs warn but proceed with defaults.

2. HIGH (dry-run filesystem mutation): default_size_fn now builds the
   index path manually (data_dir/indexes/<sanitized_id>) without calling
   index_data_dir(), which runs create_dir_all. Dry-runs no longer create
   empty per-index data dirs.

3. HIGH (colocated index deletion): add remove_data_dir_for_entry() that
   dispatches to colocated_storage::colocated_storage_dir for colocated
   entries. Previously only the global data dir was removed, leaving
   <root_path>/.trusty-search/ on disk.

4. HIGH (error counting on data-dir failure): errors += 1 now added when
   remove_data_dir_for_entry fails. Previously the failure fell through to
   print "deleted" and increment deleted.

5. HIGH (enabled flag documentation): resolved the contradiction — the
   manual --apply command does NOT require enabled=true; the enabled flag
   is documented as governing the future daemon-side sweep only. Both docs
   now agree.

6. HIGH (protected index test): prune_apply_skips_protected now injects
   AutoPruneConfig with protected_indexes via handle_prune_at's new cfg
   parameter. The test actually verifies the entry SURVIVES --apply.
   Added prune_apply_removes_eligible_preserves_protected as a combined
   scenario test.

7. MEDIUM (batch registry save): registry entries are now removed in a
   single load→filter→save (not N per-entry cycles). The race window with
   the daemon's last_queried flush is documented in the module doc.

8. MEDIUM (confirm deduplication): confirm() extracted to commands/mod.rs;
   both prune.rs and prune_orphans.rs use super::confirm().

Additionally:
- handle_prune_at now takes AutoPruneConfig as an explicit parameter so
  tests can inject protected_indexes without a real config file on disk.
- Tests split into prune_tests.rs (1500-SLOC test cap) to keep prune.rs
  well under the 500-SLOC production cap.
- New tests: prune_colocated_deletion_removes_colocated_dir,
  prune_apply_removes_eligible_preserves_protected.
